Enforce int

+++ segmentation/util/utils_postprocessing.py
@@ -1166,7 +1166,7 @@ def create_crop_masks_using_config(project_config: ModularProjectConfig, target_
     dz, dx, dy = (np.array(target_sz) / 2.0).astype(int)
 
     new_seg_fname = get_sequential_filename(old_seg_fname)
-    new_masks = zarr.open(new_seg_fname, chunks=old_masks.chunks, shape=old_masks.shape, fill_value=0)
+    new_masks = zarr.open(new_seg_fname, chunks=old_masks.chunks, shape=old_masks.shape, fill_value=0, dtype=int)
     # new_masks = zarr.zeros_like(old_masks, store=new_seg_fname)
 
     with tqdm(total=num_frames) as pbar:
@@ -1175,7 +1175,7 @@ def create_crop_masks_using_config(project_config: ModularProjectConfig, target_
             props = regionprops(labels)
             for p in tqdm(props, leave=False):
                 # Get centroids (just recalculate, unweighted)
-                label = p.label
+                label = int(p.label)
                 centroid = p.centroid
 
                 # Get bbox of the smaller size
